fixes #240 add forget password and reset password components to login-view
