+++ login-view/src/setupProxy.js
@@ -1,5 +1,6 @@
-const proxy = require('http-proxy-middleware');
+const { createProxyMiddleware } = require('http-proxy-middleware');
 
 module.exports = function(app) {
-    app.use(proxy('/oauth2/code', { target: 'https://localhost:6881', secure: false }));
+    app.use('/oauth2/code', createProxyMiddleware({ target: 'https://localhost:6881', secure: false }));
+    app.use('/portal/command', createProxyMiddleware({ target: 'https://localhost:8441', secure: false }));
 };
